server/requirements.txt
alembic==1.17.2
argon2-cffi==23.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
click==8.3.1
ecdsa==0.19.1
email-validator==2.2.0
fastapi==0.128.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
passlib[argon2]==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.1
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
